fix timer class issues (unsigned ints can't be negative..)

+++ src/timer.h
@@ -6,14 +6,14 @@
 class Timer {
 private:
     struct TimInf {
-        uint64_t rem_ms = 0;
+        int64_t rem_ms = 0;
         std::function<void()> cb;
     };
     struct TimID {
         uint64_t tim_id = 0;
         TimID& operator++() {
             tim_id++;
-            return *this;   
+            return *this;
         };
         bool operator==(const TimID& other) const {
             return tim_id == other.tim_id;
@@ -38,7 +38,7 @@ private:
     std::map<TimID, TimInf> s;
 public:
     Timer() {};
-    TimID add_timer(uint64_t ms, std::function<void()> fn) {
+    TimID add_timer(int64_t ms, std::function<void()> fn) {
         TimID x = last_id;
         ++last_id;
         s[x] = TimInf{ms, fn};
@@ -47,7 +47,7 @@ public:
     void remove(TimID xid) {
         s.erase(xid);
     };
-    void advance(int ms_fact) {
+    void advance(int64_t ms_fact) {
         for (auto i = s.begin(); i != s.end() ; ) {
             i->second.rem_ms -= ms_fact;
             if (i->second.rem_ms <= 0) {
